updated wall stopping mechanics to work better including adding gravity to all walls
