+++ Assets/Material.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c29f09fe3de14b1478cb9b39e58496f2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
