+++ Assets/Objects/Player.prefab
@@ -15,7 +15,6 @@ GameObject:
   - component: {fileID: 5787757993737344035}
   - component: {fileID: 2829820299888922210}
   - component: {fileID: 1800262915088092133}
-  - component: {fileID: 5708471897185805801}
   - component: {fileID: 6646864741515115909}
   - component: {fileID: 515385037024534229}
   m_Layer: 0
@@ -208,41 +207,6 @@ Tilemap:
     e31: 0
     e32: 0
     e33: 1
---- !u!70 &5708471897185805801
-CapsuleCollider2D:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 961402715714136404}
-  m_Enabled: 0
-  m_Density: 1
-  m_Material: {fileID: 0}
-  m_IncludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_ExcludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_LayerOverridePriority: 0
-  m_ForceSendLayers:
-    serializedVersion: 2
-    m_Bits: 4294967295
-  m_ForceReceiveLayers:
-    serializedVersion: 2
-    m_Bits: 4294967295
-  m_ContactCaptureLayers:
-    serializedVersion: 2
-    m_Bits: 4294967295
-  m_CallbackLayers:
-    serializedVersion: 2
-    m_Bits: 4294967295
-  m_IsTrigger: 0
-  m_UsedByEffector: 0
-  m_UsedByComposite: 0
-  m_Offset: {x: -0.14801168, y: -0.0612629}
-  m_Size: {x: 0.77496326, y: 0.94493526}
-  m_Direction: 0
 --- !u!82 &6646864741515115909
 AudioSource:
   m_ObjectHideFlags: 0
