+++ Assets/Objects/SlideWall.prefab
@@ -10,10 +10,10 @@ GameObject:
   m_Component:
   - component: {fileID: 622371382286947111}
   - component: {fileID: 60006535751492107}
-  - component: {fileID: 5135162769575836848}
   - component: {fileID: 5886912171426859905}
-  - component: {fileID: 463687148284121848}
-  - component: {fileID: 4282817068625861387}
+  - component: {fileID: 4332905782556579567}
+  - component: {fileID: 5473416651320500331}
+  - component: {fileID: 125947664339434096}
   m_Layer: 0
   m_Name: SlideWall
   m_TagString: MapObject
@@ -88,7 +88,19 @@ SpriteRenderer:
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
---- !u!61 &5135162769575836848
+--- !u!114 &5886912171426859905
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6853101624086219989}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 0d1eb2b77763a3a40bcee5cdc6f5e6c7, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!61 &4332905782556579567
 BoxCollider2D:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -97,7 +109,7 @@ BoxCollider2D:
   m_GameObject: {fileID: 6853101624086219989}
   m_Enabled: 1
   m_Density: 1
-  m_Material: {fileID: 0}
+  m_Material: {fileID: 6200000, guid: dfa9033817e1a724f8ef1a7a26e7ff37, type: 2}
   m_IncludeLayers:
     serializedVersion: 2
     m_Bits: 0
@@ -120,7 +132,7 @@ BoxCollider2D:
   m_IsTrigger: 0
   m_UsedByEffector: 0
   m_UsedByComposite: 0
-  m_Offset: {x: -0.00090277195, y: 0.10080242}
+  m_Offset: {x: 0, y: 0}
   m_SpriteTilingProperty:
     border: {x: 0, y: 0, z: 0, w: 0}
     pivot: {x: 0.5, y: 0.5}
@@ -131,21 +143,9 @@ BoxCollider2D:
     adaptiveTiling: 0
   m_AutoTiling: 0
   serializedVersion: 2
-  m_Size: {x: 0.55760884, y: 4.798395}
+  m_Size: {x: 0.55, y: 5}
   m_EdgeRadius: 0
---- !u!114 &5886912171426859905
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6853101624086219989}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0d1eb2b77763a3a40bcee5cdc6f5e6c7, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
---- !u!61 &463687148284121848
+--- !u!61 &5473416651320500331
 BoxCollider2D:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -177,7 +177,7 @@ BoxCollider2D:
   m_IsTrigger: 1
   m_UsedByEffector: 0
   m_UsedByComposite: 0
-  m_Offset: {x: -0.00090289116, y: 0.10080171}
+  m_Offset: {x: 0, y: 0}
   m_SpriteTilingProperty:
     border: {x: 0, y: 0, z: 0, w: 0}
     pivot: {x: 0.5, y: 0.5}
@@ -188,9 +188,9 @@ BoxCollider2D:
     adaptiveTiling: 0
   m_AutoTiling: 0
   serializedVersion: 2
-  m_Size: {x: 0.5539975, y: 4.7983966}
+  m_Size: {x: 0.55, y: 5}
   m_EdgeRadius: 0
---- !u!114 &4282817068625861387
+--- !u!114 &125947664339434096
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -199,7 +199,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 6853101624086219989}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e7546bb7086a22c4fbfde2638d784aa4, type: 3}
+  m_Script: {fileID: 11500000, guid: 8adea71ea3c447a488874332864fa572, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  GravityValue: 0.2
+  gravityValue: 0.2
