+++ Assets/Objects/SmallBlock.prefab
@@ -12,8 +12,8 @@ GameObject:
   - component: {fileID: 5849160294308586115}
   - component: {fileID: 7335354203521496574}
   - component: {fileID: -5607395088704061429}
-  - component: {fileID: 8960730265113829756}
   - component: {fileID: 8292184434631423054}
+  - component: {fileID: 342717064081035911}
   m_Layer: 0
   m_Name: SmallBlock
   m_TagString: Untagged
@@ -97,7 +97,7 @@ BoxCollider2D:
   m_GameObject: {fileID: 3584316300872367206}
   m_Enabled: 1
   m_Density: 1
-  m_Material: {fileID: 0}
+  m_Material: {fileID: 6200000, guid: dfa9033817e1a724f8ef1a7a26e7ff37, type: 2}
   m_IncludeLayers:
     serializedVersion: 2
     m_Bits: 0
@@ -120,7 +120,7 @@ BoxCollider2D:
   m_IsTrigger: 0
   m_UsedByEffector: 0
   m_UsedByComposite: 0
-  m_Offset: {x: 0, y: 0.072849154}
+  m_Offset: {x: 0, y: 0}
   m_SpriteTilingProperty:
     border: {x: 0, y: 0, z: 0, w: 0}
     pivot: {x: 0.5, y: 0.5}
@@ -131,7 +131,7 @@ BoxCollider2D:
     adaptiveTiling: 0
   m_AutoTiling: 0
   serializedVersion: 2
-  m_Size: {x: 1, y: 0.8543017}
+  m_Size: {x: 1, y: 1}
   m_EdgeRadius: 0
 --- !u!61 &-5607395088704061429
 BoxCollider2D:
@@ -165,7 +165,7 @@ BoxCollider2D:
   m_IsTrigger: 1
   m_UsedByEffector: 0
   m_UsedByComposite: 0
-  m_Offset: {x: 0, y: 0.072849154}
+  m_Offset: {x: 0, y: 0}
   m_SpriteTilingProperty:
     border: {x: 0, y: 0, z: 0, w: 0}
     pivot: {x: 0.5, y: 0.5}
@@ -176,20 +176,8 @@ BoxCollider2D:
     adaptiveTiling: 0
   m_AutoTiling: 0
   serializedVersion: 2
-  m_Size: {x: 1, y: 0.8543017}
+  m_Size: {x: 1, y: 1}
   m_EdgeRadius: 0
---- !u!114 &8960730265113829756
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3584316300872367206}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e7546bb7086a22c4fbfde2638d784aa4, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!114 &8292184434631423054
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -202,3 +190,16 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 0d1eb2b77763a3a40bcee5cdc6f5e6c7, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &342717064081035911
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3584316300872367206}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8adea71ea3c447a488874332864fa572, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  gravityValue: 0.01
