+++ Assets/Scripts/PlayerWallTouch.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e7546bb7086a22c4fbfde2638d784aa4
+guid: 8adea71ea3c447a488874332864fa572
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
